Use Mega key to control DIAG SENSE.
This way, if you reset while Mega is pressed, with most ROM versions you
will end up in the monitor before memory is wiped. The key was chosen
because it is close to the reset button. The diagnostic sense input is
only used while booting, so the fact that the same key is used for other
purposes as well should not matter. (If it does, another key can be
chosen)