Copy hsync/vsync from input
instead of making it the same as the blanking signals.
The way the timing for the blanking is calculated is supposed to be a
superset of the syncing time. It is not so nice to have them the same,
even if the IMDH things in the video pipeline don't seem to care about
that much.