More clock domain crossing changes.
img_mounted is a brief (strobe) signal, so if we send it unchanged
through the _sync module, and the clock frequencies are unlucky, we
could miss the strobe. So we stretch it first.

drv_type is used in both clock domains, so let's input it again as
clk_sys signal and do the crossing internally, so we have both versions
available.