Disable screen snow for 8296.
The current implementation gives too much snow (the whole screen full)
because writes to ROM addresses can write to the ram "under" it, which
is the same ram block as the video ram. As the cpu code runs from ROM,
these addresses are accessed pretty much all the time.

The combination didn't occur in real hardware anyway, so just disable it.

In theory perhaps the snow could be avoided for ROM read cycles which
effectively do not read the RAM (in real hardware, CAS would probably not
be asserted for that), but that would mean reorganizing a lot of logic
and I don't think it is worth it for now.

Do take the opportunity to add more comments and add extra wires for the
several meanings of pref_eoi_blanks, for clarity.