Implement reading from an open bus.
This makes that (usually!) reading from open space returns the high byte
of the address. Reason: that's the last thing read from memory, either as
an zero page pointer address, or from the instruction.

Exceptions exist for indexed addressing with a page crossing, where there
is a dummy cycle which can cause a different value on the bus.

Too bad that our ROMs at $9xxx, $Axxx (and maybe $Bxxx) are not really
empty, but are pre-filled with "seeming empty space". But it is effective
at least at $E80x, $8800-$8FFF, and below $8000 if an 8K or 16K model
is selected.

While there, also loop back the cpu data output (when it's a write
cycle) back to its input. This is easy to put into the same mux.
The cpu needs this for some undocumented opcodes:
--   TAKE CARE you route the dout signal back to the din signal while R_W_n='0',
--   otherwise some undocumented opcodes won't work correctly.