Simplify vram_addr_vid_[02].
They were essentially the same, apart from how many significant bits
from video_addr they used, depending on the model (8296 or earlier).
If we make the distinction in the definition of video_addr itself,
we need only one of the two. video_addr was simply a copy of crtc_ma;
now we mask in or out the higher bits depending on the model of PET.