Improve Data bus latching towards 6809.
We want to use the value at the rising edge of "enable" (same as the
6502). Previous code had a delay and used the value from 1 fast clock
earlier. In practice that works too, since the video circuitry (which
plays with the address bus of the vram, which is what makes this latch
desirable) is careful to stay clear of cycles close to "enable". But
it's better to do it correctly.